requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
drf-spectacular
